requirements.txt
telethon==1.41.2
python-dotenv==1.2.1
aiogram==3.22.0